move report parsing related code to the parse cmd
the PlistToPlaintextFormatter and the skip_report
function is only used at the parse command
