+++ codechecker_common/plist_parser.py
@@ -33,10 +33,8 @@ from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
 
-from collections import defaultdict
 import imp
 import io
-import math
 import os
 import sys
 import traceback
@@ -44,12 +42,8 @@ from plistlib import PlistParser, writePlist, writePlistToString, \
     readPlistFromString
 from xml.parsers.expat import ExpatError
 
-from codechecker_common import util
 from codechecker_common.logger import get_logger
-from codechecker_common.report import Report, generate_report_hash, \
-    get_report_path_hash
-from codechecker_common.source_code_comment_handler import \
-    SourceCodeCommentHandler, skip_suppress_status
+from codechecker_common.report import Report, generate_report_hash
 
 LOG = get_logger('report')
 
@@ -340,278 +334,3 @@ def skip_report_from_plist(plist_file, skip_handler):
         plist.seek(0)
         plist.write(new_plist_content)
         plist.truncate()
-
-
-def skip_report(report_hash, source_file, report_line, checker_name,
-                src_comment_handler=None):
-    """
-    Returns True if the report was suppressed in the source code, otherwise
-    False.
-    """
-    bug = {'hash_value': report_hash, 'file_path': source_file}
-    if src_comment_handler and src_comment_handler.get_suppressed(bug):
-        LOG.debug("Suppressed by suppress file: %s:%s [%s] %s", source_file,
-                  report_line, checker_name, report_hash)
-        return True
-
-    sc_handler = SourceCodeCommentHandler()
-
-    # Check for source code comment.
-    src_comment_data = sc_handler.filter_source_line_comments(
-        source_file,
-        report_line,
-        checker_name)
-
-    if len(src_comment_data) == 1:
-        status = src_comment_data[0]['status']
-
-        LOG.debug("Suppressed by source code comment.")
-        if src_comment_handler:
-            file_name = os.path.basename(source_file)
-            message = src_comment_data[0]['message']
-            src_comment_handler.store_suppress_bug_id(
-                report_hash,
-                file_name,
-                message,
-                status)
-
-        if skip_suppress_status(status):
-            return True
-
-    elif len(src_comment_data) > 1:
-        LOG.warning("Multiple source code comment can be found "
-                    "for '%s' checker in '%s' at line %d. "
-                    "This bug will not be suppressed!",
-                    checker_name, source_file, report_line)
-    return False
-
-
-class PlistToPlaintextFormatter(object):
-    """
-    Parse and format plist reports to a more human readable format.
-    """
-
-    def __init__(self,
-                 src_comment_handler,
-                 skip_handler,
-                 severity_map,
-                 processed_path_hashes,
-                 trim_path_prefixes,
-                 analyzer_type="clangsa"):
-
-        self.__analyzer_type = analyzer_type
-        self.__severity_map = severity_map
-        self.print_steps = False
-        self.src_comment_handler = src_comment_handler
-        self.skiplist_handler = skip_handler
-        self._processed_path_hashes = processed_path_hashes
-        self._trim_path_prefixes = trim_path_prefixes
-
-    @staticmethod
-    def __format_location(event, source_file):
-        loc = event['location']
-        line = util.get_line(source_file, loc['line'])
-        if line == '':
-            return line
-
-        marker_line = line[0:(loc['col'] - 1)]
-        marker_line = ' ' * (len(marker_line) + marker_line.count('\t'))
-        return '%s%s^' % (line.replace('\t', '  '), marker_line)
-
-    @staticmethod
-    def __format_bug_event(name, severity, event, source_file):
-
-        loc = event['location']
-        fname = os.path.basename(source_file)
-        if name:
-            return '[%s] %s:%d:%d: %s [%s]' % (severity,
-                                               source_file,
-                                               loc['line'],
-                                               loc['col'],
-                                               event['message'],
-                                               name)
-        else:
-            return '%s:%d:%d: %s' % (fname,
-                                     loc['line'],
-                                     loc['col'],
-                                     event['message'])
-
-    @staticmethod
-    def __format_bug_note(note, source_file):
-        """
-        Format bug notes.
-        """
-        loc = note['location']
-        file_name = os.path.basename(source_file)
-        return '%s:%d:%d: %s' % (file_name,
-                                 loc['line'],
-                                 loc['col'],
-                                 note['message'])
-
-    @staticmethod
-    def __format_macro_expansion(macro, source_file):
-        """
-        Format macro expansions.
-        """
-        loc = macro['location']
-        file_name = os.path.basename(source_file)
-        return "%s:%d:%d: Macro '%s' expanded to '%s'" % (file_name,
-                                                          loc['line'],
-                                                          loc['col'],
-                                                          macro['name'],
-                                                          macro['expansion'])
-
-    @staticmethod
-    def parse(plist_file):
-        """
-        Parse a plist report file.
-        Returns:
-            - list of source files
-            - list of reports (type Report)
-        """
-        files, reports = [], []
-        try:
-            files, reports = parse_plist_file(plist_file)
-        except Exception as ex:
-            traceback.print_stack()
-            LOG.error('The generated plist is not valid!')
-            LOG.error(ex)
-        finally:
-            return files, reports
-
-    def write(self, file_report_map, output=sys.stdout):
-        """
-        Format an already parsed plist report file to a more
-        human readable format.
-        The formatted text is written to the output.
-        During writing the output statistics are collected.
-
-        Write out the bugs to the output and collect report statistics.
-        """
-
-        severity_stats = defaultdict(int)
-        file_stats = defaultdict(int)
-        report_count = defaultdict(int)
-
-        for file_path in sorted(file_report_map,
-                                key=lambda key: len(file_report_map[key])):
-
-            non_suppressed = 0
-            sorted_reports = sorted(file_report_map[file_path],
-                                    key=lambda r: r.main['location']['line'])
-
-            for report in sorted_reports:
-                path_hash = get_report_path_hash(report, report.files)
-                if path_hash in self._processed_path_hashes:
-                    LOG.debug("Not showing report because it is a "
-                              "deduplication of an already processed report!")
-                    LOG.debug("Path hash: %s", path_hash)
-                    LOG.debug(report)
-                    continue
-
-                self._processed_path_hashes.add(path_hash)
-
-                events = [i for i in report.bug_path
-                          if i.get('kind') == 'event']
-                f_path = report.files[events[-1]['location']['file']]
-                if self.skiplist_handler and \
-                        self.skiplist_handler.should_skip(f_path):
-                    LOG.debug("Skipped report in '%s'", f_path)
-                    LOG.debug(report)
-                    continue
-
-                last_report_event = report.bug_path[-1]
-                source_file = \
-                    report.files[last_report_event['location']['file']]
-                trimmed_source_file = \
-                    util.trim_path_prefixes(source_file,
-                                            self._trim_path_prefixes)
-
-                report_line = last_report_event['location']['line']
-                report_hash = \
-                    report.main['issue_hash_content_of_line_in_context']
-                checker_name = report.main['check_name']
-
-                if skip_report(report_hash, source_file, report_line,
-                               checker_name, self.src_comment_handler):
-                    continue
-
-                file_stats[f_path] += 1
-                severity = self.__severity_map.get(checker_name)
-                severity_stats[severity] += 1
-                report_count["report_count"] += 1
-
-                output.write(self.__format_bug_event(checker_name,
-                                                     severity,
-                                                     last_report_event,
-                                                     trimmed_source_file))
-                output.write('\n')
-                output.write(self.__format_location(last_report_event,
-                                                    source_file))
-                output.write('\n')
-
-                if self.print_steps:
-                    output.write('  Report hash: ' + report_hash + '\n')
-
-                    # Print out macros.
-                    macros = report.macro_expansions
-                    if macros:
-                        output.write('  Macro expansions:\n')
-
-                        index_format = '    %%%dd, ' % \
-                                       int(math.floor(
-                                           math.log10(len(macros))) + 1)
-
-                        for index, macro in enumerate(macros):
-                            output.write(index_format % (index + 1))
-                            source = report.files[
-                                macro['location']['file']]
-                            output.write(self.__format_macro_expansion(macro,
-                                                                       source))
-                            output.write('\n')
-
-                    # Print out notes.
-                    notes = report.notes
-                    if notes:
-                        output.write('  Notes:\n')
-
-                        index_format = '    %%%dd, ' % \
-                                       int(math.floor(
-                                           math.log10(len(notes))) + 1)
-
-                        for index, note in enumerate(notes):
-                            output.write(index_format % (index + 1))
-                            source_file = report.files[
-                                note['location']['file']]
-                            output.write(self.__format_bug_note(note,
-                                                                source_file))
-                            output.write('\n')
-
-                    output.write('  Steps:\n')
-
-                    index_format = '    %%%dd, ' % \
-                                   int(math.floor(math.log10(len(events))) + 1)
-
-                    for index, event in enumerate(events):
-                        output.write(index_format % (index + 1))
-                        source_file = report.files[event['location']['file']]
-                        output.write(
-                            self.__format_bug_event(None,
-                                                    None,
-                                                    event,
-                                                    trimmed_source_file))
-                        output.write('\n')
-                output.write('\n')
-
-                non_suppressed += 1
-
-            base_file = os.path.basename(file_path)
-            if non_suppressed == 0:
-                output.write('Found no defects in %s\n' % base_file)
-            else:
-                output.write('Found %d defect(s) in %s\n\n' %
-                             (non_suppressed, base_file))
-
-        return {"severity": severity_stats,
-                "files":  file_stats,
-                "reports": report_count}
